UCL v2.00
The latest version of NeuroMatic (2.00) is now available for downloading at:

http://www.neuromatic.thinkrandom.com

Features of the latest NeuroMatic include:

    -works with Igor 5 and 6
    -a new "Fit" tab for semi-automated curve fitting
    -several new Stats functions, including Onset and FWHM measurements
    -misc bug fixes

There is also a new version of Nclamp, the acquisition aspect of
NeuroMatic, which can be downloaded at:

http://www.physiol.ucl.ac.uk/research/silver_a/nclamp/

Features of the latest Nclamp include:

    -a simplified user interface
    -global board configurations that are now saved independent of
stimulus protocols
    -works with the latest Igor Nidaq MX XOP
    -improved online "Stats" functionality
    -new online "Spike" detection for creating raster plots while recording
    -misc bug fixes

NIDAQ users are recommended to use the latest MX drivers and switch to
using Nclamp 2.00.

More details are given on the above webpages.

Questions can be sent to Jason@ThinkRandom.com

Thanks!